generate pg-mem conection an base service for CRUD

+++ apps/edwin-guamushig_repos/src/base/base-app.service.ts
@@ -7,16 +7,16 @@ import { Repository } from 'typeorm';
 import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';
 
 @Injectable()
-export class BaseAppService<Entity> {
+export class BaseAppService<Entity, CreateDto> {
   protected readonly _repository: Repository<Entity>;
 
   constructor(_repository: Repository<Entity>) {
     this._repository = _repository;
   }
 
-  async create(data: Entity): Promise<Entity> {
+  async create(data: CreateDto): Promise<Entity> {
     try {
-      return this._repository.save(data);
+      return this._repository.save(data as any);
     } catch (e) {
       throw new InternalServerErrorException('Error creating');
     }
@@ -26,12 +26,11 @@ export class BaseAppService<Entity> {
     id: number,
     updateData: QueryDeepPartialEntity<Entity>,
   ): Promise<void> {
+    const data = await this._repository.findOneById(id);
+    if (!data) {
+      throw new NotFoundException(`Registry with id ${id} not found`);
+    }
     try {
-      const data = await this._repository.findOneById(id);
-      if (!data) {
-        throw new NotFoundException(`Registry with id ${id} not found`);
-      }
-
       await this._repository.update(id, { ...updateData });
     } catch (e) {
       throw new InternalServerErrorException('Error updating');
@@ -48,19 +47,20 @@ export class BaseAppService<Entity> {
 
   async findById(id: number): Promise<Entity> {
     try {
-      const data = await this._repository.findOneById(id);
-      if (!data) {
-        throw new NotFoundException(`Registry with id ${id} not found`);
-      }
-      return data;
+      return this._repository.findOneById(id);
     } catch (e) {
       throw new InternalServerErrorException('Error finding by id');
     }
   }
 
   async delete(id: number): Promise<void> {
+    const registro = await this._repository.findOneById(id);
+
+    if (!registro) {
+      throw new NotFoundException(`Registry with id ${id} not found`);
+    }
+
     try {
-      const registro = await this._repository.findOneById(id);
       await this._repository.remove(registro);
     } catch (e) {
       throw new InternalServerErrorException('Error deleting');
